Moved Google Calendar functions to a separate extension
