+++ lib/pages/timetable/service/uni_event_provider.dart
@@ -26,6 +26,7 @@ import 'package:googleapis_auth/auth_io.dart';
 import 'package:rrule/rrule.dart';
 import 'package:time_machine/time_machine.dart';
 import 'package:timetable/timetable.dart';
+import 'google_calendar_services.dart';
 
 extension PeriodExtension on Period {
   static Period fromJSON(Map<String, dynamic> json) {
@@ -311,56 +312,6 @@ class UniEventProvider extends EventProvider<UniEventInstance>
     return stream.map((events) => events.expand((i) => i).toList());
   }
 
-  g_cal.Event convertEvent(UniEvent uniEvent) {
-    final g_cal.Event _gCalEvent = g_cal.Event();
-
-    final g_cal.EventDateTime start = g_cal.EventDateTime();
-    final DateTime startDateTime = uniEvent.start.toDateTimeLocal();
-
-    start
-      //..timeZone = startDateTime.timeZoneName
-      ..timeZone = 'Europe/Bucharest' // EET
-      ..dateTime = startDateTime;
-    final Period eventPeriod = uniEvent.duration;
-    final Duration duration =
-        Duration(hours: eventPeriod.hours, minutes: eventPeriod.minutes);
-
-    final g_cal.EventDateTime end = g_cal.EventDateTime();
-    final DateTime endDateTime = startDateTime.add(duration);
-    end
-      //..timeZone = endDateTime.timeZoneName
-      ..timeZone = 'Europe/Bucharest'
-      ..dateTime = endDateTime;
-
-    // TODO(bogpie): Recreate timezone from existing information
-    // TODO(bogpie): Require user to input how many minutes before a notification from GCal (including the "no notification" option)
-
-    final ClassHeader classHeader = uniEvent.classHeader;
-    _gCalEvent
-      ..start = start
-      ..end = end
-      ..summary = classHeader.acronym
-      ..colorId = (uniEvent.type.googleCalendarColor.index).toString()
-/*      ..description =  eventInstance.title ??
-          eventInstance.mainEvent.type
-              .toLocalizedString(context)*/
-      // TODO(bogpie): Use a relevant description, like type of class + lecturer
-
-      ..location = uniEvent.location;
-
-    if (uniEvent is RecurringUniEvent) {
-      _gCalEvent.recurrence = <String>[];
-
-      String newRruleString = uniEvent.newRrule.toString();
-      newRruleString =
-          newRruleString.replaceAll(RegExp(r'T000000'), 'T000000Z');
-      newRruleString = newRruleString.replaceAll(RegExp(r'DAILY'), 'YEARLY');
-      print(newRruleString);
-      _gCalEvent.recurrence.add(newRruleString);
-    }
-
-    return _gCalEvent;
-  }
 
   Future<void> exportToGoogleCalendar(TimetablePage widget) async {
     final Stream<List<UniEvent>> eventsStream = _events;
@@ -380,81 +331,6 @@ class UniEventProvider extends EventProvider<UniEventInstance>
     ///RRULE:FREQ=WEEKLY;UNTIL=20210605T000000;INTERVAL=1;BYDAY=WE doesn't work
 
     await insertGoogleEvents(_gCalEvents);
-  }
-
-  void prompt(String url) async {
-    print('Please go to the following URL and grant access:');
-    print('  => $url');
-    print('');
-
-    await FlutterWebBrowser.openWebPage(url: url);
-  }
-
-  Future<void> insertGoogleEvents(List<g_cal.Event> _gCalEvents) async {
-    await clientViaUserConsent(
-            GoogleApiHelper.credentials, GoogleApiHelper.scopes, prompt)
-        .then(
-      //(AuthClient client) async {
-      (AutoRefreshingAuthClient client) async {
-        // TODO(bogpie): Remember if a user already gave access to his Google Calendar
-        // TODO(bogpie): Automatically close browser
-
-        final g_cal.CalendarApi calendarApi = g_cal.CalendarApi(client);
-        final g_cal.Calendar calendar = g_cal.Calendar();
-        g_cal.CalendarList calendarListNonIterable;
-        try {
-          calendarListNonIterable = await calendarApi.calendarList.list();
-        } catch (e) {
-          print('Error $e in getting calendars as a list');
-          return;
-        }
-        final List<g_cal.CalendarListEntry> calendarList =
-            calendarListNonIterable.items;
-        for (final g_cal.CalendarListEntry calendar in calendarList) {
-          if (calendar.summary == 'ACS UPB Mobile') {
-            try {
-              await calendarApi.calendars.delete(calendar.id);
-            } catch (e) {
-              print('Error $e in deleting calendar');
-            }
-            break;
-          }
-        }
-        // ignore: cascade_invocations
-        calendar
-          ..timeZone = 'Europe/Bucharest'
-          ..summary = 'ACS UPB Mobile'
-          ..description = 'Timetable imported from ACS UPB Mobile';
-
-        final g_cal.Calendar returnedCalendar =
-            await calendarApi.calendars.insert(calendar);
-
-        if (returnedCalendar is g_cal.Calendar) {
-          final String calendarId = returnedCalendar.id;
-          for (final g_cal.Event event in _gCalEvents) {
-            try {
-              await calendarApi.events.insert(event, calendarId).then(
-                (value) {
-                  print('Added event status: ${value.status}');
-                  if (value.status == 'confirmed') {
-                    print(
-                        'Event named ${event.myToString()} added in Google Calendar'); //log
-                  } else {
-                    print(
-                        'Unable to add event named ${event.myToString()} in Google Calendar');
-                  }
-                },
-              );
-            } catch (e) {
-              print('Error creating event $e');
-            }
-          }
-        }
-      },
-      onError: (dynamic e) {
-        print('Error <$e> when asking for user\'s consent');
-      },
-    );
   }
 
   @override
